Train pipeline

+++ marionet/dataset/dataset.py
@@ -104,4 +104,4 @@ class MarioNetDataset(Dataset):
         :param image: Image in numpy format, C x H x W.
         """
         image = torch.tensor(image)
-        return self.transforms(image)
+        return self.transforms(image).float()
